pyproject.toml
[tool.poetry]
name = "jordantracker"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.0.3"
extract-msg = "^0.42.1"
html2text = "2020.1.16"
slackclient = "^2.9.4"
pytest = "^7.4.0"

[tool.poetry.scripts]
jordantracker = "jordantracker.jordantracker:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
